+++ R/plot.R
@@ -116,47 +116,50 @@ setMethod(
     }
 )
 
-init_plot_compartments <- function(x, compartments) {
-    ## Determine the compartments to include in the plot
-    if (is.null(compartments))
-        compartments <- rownames(x@S)
-    if (!(all(compartments %in% rownames(x@S))))
-        stop("'compartments' must exist in the model.", call. = FALSE)
-    match(compartments, rownames(x@S))
+init_plot_node_index <- function(model, index) {
+    index <- check_node_index_argument(model, index)
+    if (is.null(index))
+        index <- seq_len(n_nodes(model))
+    index
 }
 
-init_plot_node <- function(x, node) {
-    node <- check_node_index_argument(x, node)
-    if (is.null(node))
-        node <- seq_len(n_nodes(x))
-    node
-}
-
-init_plot_line_type <- function(lty, compartments, m) {
-    if (is.null(lty)) {
-        lty <- seq_len(length(compartments))
+init_plot_line_type <- function(lty, compartments, each) {
+    if (is.null(compartments)) {
+        n <- 1
     } else {
-        lty <- rep(lty, length.out = length(compartments))
+        n <- length(compartments)
     }
-    rep(lty, length.out = dim(m)[1])
+
+    if (is.null(lty)) {
+        lty <- seq_len(n)
+    } else {
+        lty <- rep(lty, length.out = n)
+    }
+    rep(lty, each = each)
 }
 
-init_plot_color <- function(col, compartments, m) {
+init_plot_color <- function(col, compartments, each) {
+    if (is.null(compartments)) {
+        n <- 1
+    } else {
+        n <- length(compartments)
+    }
+
     if (is.null(col)) {
-        if (length(compartments) > 9) {
-            col <- rainbow(length(compartments))
-        } else if (length(compartments) > 1) {
+        if (n > 9) {
+            col <- rainbow(n)
+        } else if (n > 1) {
             col <- rep(c("#e41a1c", "#377eb8", "#4daf4a",
                          "#984ea3", "#ff7f00", "#ffff33",
                          "#a65628", "#f781bf", "#999999"),
-                       length.out = length(compartments))
+                       length.out = n)
         } else {
             col <- "black"
         }
     } else {
-        col <- rep(col, length.out = length(compartments))
+        col <- rep(col, length.out = n)
     }
-    rep(col, length.out = dim(m)[1])
+    rep(col, each = each)
 }
 
 init_plot_type <- function(type) {
@@ -171,16 +174,219 @@ init_plot_line_width <- function(lwd) {
     lwd
 }
 
+init_plot_range <- function(range) {
+    if (identical(range, FALSE))
+        return(range)
+
+    if (any(!is.numeric(range),
+            !identical(length(range), 1L),
+            range < 0, range > 1)) {
+        stop("'range' must be FALSE or a value between 0 and 1.",
+             call. = FALSE)
+    }
+
+    (1 - range) / 2
+}
+
+init_plot_prevalence_data <- function(model, compartments,
+                                      level, index, range) {
+    index <- init_plot_node_index(model, index)
+    range <- init_plot_range(range)
+
+    ## Create a matrix with one row for each line in the plot.
+    if (level < 3) {
+        y <- prevalence(model, compartments, level, index, "matrix")
+        lower <- NULL
+        upper <- NULL
+        each <- 1
+    } else if (identical(range, FALSE)) {
+        y <- prevalence(model, compartments, level, index, "matrix")
+        lower <- NULL
+        upper <- NULL
+        each <- length(index)
+    } else {
+        y <- apply(
+            prevalence(model, compartments, level, index, "matrix"),
+            2, quantile, probs = c(range, 0.5, 1 - range))
+
+        ## Matrices for quantile ranges and median.
+        j <- seq_len(ncol(y))
+        lower <- y[1, j, drop = FALSE]
+        upper <- y[3, j, drop = FALSE]
+        y <- y[2, j, drop = FALSE]
+        each <- 1
+    }
+
+    list(lower        = lower,
+         y            = y,
+         upper        = upper,
+         each         = each,
+         compartments = NULL)
+}
+
+init_plot_trajectory_data <- function(model, compartments, index, range) {
+    index <- init_plot_node_index(model, index)
+    range <- init_plot_range(range)
+
+    ## Create a matrix with one row for each line in the plot.
+    y <- list()
+    for (j in seq_len(length(compartments$rhs))) {
+        for (compartment in names(compartments$rhs[[j]])) {
+            if (identical(range, FALSE)) {
+                y[[length(y) + 1]] <-
+                    trajectory(model, compartment, index, "matrix")
+            } else {
+                y[[length(y) + 1]] <- apply(
+                    trajectory(model, compartment, index, "matrix"),
+                    2, quantile, probs = c(range, 0.5, 1 - range))
+            }
+
+            names(y)[length(y)] <- compartment
+        }
+    }
+
+    compartments <- names(y)
+
+    if (identical(range, FALSE)) {
+        lower <- NULL
+        upper <- NULL
+        ## Combine matrices for each comparment.
+        y <- do.call("rbind", y)
+        each <- length(index)
+    } else {
+        ## Matrices for quantile ranges and median.
+        lower <- do.call("rbind", lapply(y, function(x) x[1, ]))
+        upper <- do.call("rbind", lapply(y, function(x) x[3, ]))
+        y <- do.call("rbind", lapply(y, function(x) x[2, ]))
+        each <- 1
+    }
+
+    list(lower        = lower,
+         y            = y,
+         upper        = upper,
+         each         = each,
+         compartments = compartments)
+}
+
+##' Determine if the 'compartments' expression contains a lhs.
+##' @noRd
+compartments_has_lhs <- function(compartments) {
+    if (is(compartments, "formula")) {
+        compartments <- as.character(compartments)
+        if (identical(length(compartments), 3L))
+            return(TRUE)
+    }
+
+    FALSE
+}
+
+init_plot_argv <- function(model, compartments, pd, ...) {
+    argv <- list(...)
+
+    if (is.null(argv$ylab)) {
+        if (isTRUE(compartments_has_lhs(compartments))) {
+            argv$ylab <- deparse(compartments)
+        } else {
+            argv$ylab <- "Value"
+        }
+    }
+
+    argv$type <- init_plot_type(argv$type)
+    argv$lwd <- init_plot_line_width(argv$lwd)
+
+    ## Settings for the y-axis.
+    if (is.null(argv$ylim)) {
+        if (is.null(pd$upper)) {
+            argv$ylim <- c(0, max(pd$y))
+        } else {
+            argv$ylim <- c(0, max(pd$upper))
+        }
+    }
+
+    ## Settings for the x-axis
+    if (is.null(names(model@tspan))) {
+        argv$x <- model@tspan
+        if (is.null(argv$xlab))
+            argv$xlab <- "Time"
+    } else {
+        argv$x <- as.Date(names(model@tspan))
+        if (is.null(argv$xlab))
+            argv$xlab <- "Date"
+    }
+
+    argv
+}
+
+plot_data <- function(pd, argv, lty, col) {
+    if (is.null(pd$compartments)) {
+        savepar <- par(mar = c(2, 4, 1, 1), oma = c(2, 1, 0, 0), xpd = TRUE)
+    } else {
+        savepar <- par(mar = c(2, 4, 1, 1), oma = c(4, 1, 0, 0), xpd = TRUE)
+    }
+    on.exit(par(savepar))
+
+    ## Plot lines
+    for (i in seq_len(dim(pd$y)[1])) {
+        argv$y <- pd$y[i, ]
+        argv$col <- col[i]
+        argv$lty <- lty[i]
+
+        if (i == 1) {
+            do.call(plot, argv)
+            title(xlab = argv$xlab, outer = TRUE, line = 0)
+        } else {
+            do.call(lines, argv)
+        }
+
+        if (!is.null(pd$lower) && !is.null(pd$upper)) {
+            polygon(x = c(argv$x, rev(argv$x)),
+                    y = c(pd$upper[i, ], rev(pd$lower[i, ])),
+                    col = adjustcolor(col[i], alpha.f = 0.1),
+                    border = NA)
+        }
+    }
+
+    ## Add the legend below plot. The default legend is the names
+    ## of the compartments.
+    if (!is.null(pd$compartments)) {
+        par(fig = c(0, 1, 0, 1), oma = c(0, 0, 0, 0),
+            mar = c(0, 0, 0, 0), new = TRUE)
+        plot(0, 0, type = "n", bty = "n", xaxt = "n", yaxt = "n")
+        legend("bottom", inset = c(0, 0),
+               lty = lty[seq_len(length(pd$compartments))],
+               col = col[seq_len(length(pd$compartments))],
+               bty = "n", horiz = TRUE, legend = pd$compartments,
+               lwd = argv$lwd)
+    }
+}
+
 ##' Display the outcome from a simulated trajectory
 ##'
 ##' Plot either the median and the quantile range of the counts in all
 ##' nodes, or plot the counts in specified nodes.
 ##' @param x The \code{model} to plot.
-##' @param compartments Character vector with the compartments in the
-##'     model to include in the plot. Default is \code{NULL}
-##'     i.e. include all compartments in the model.
-##' @param node Indices specifying the nodes to include when plotting
-##'     data. Plot one line for each node. Default (\code{node =
+##' @param y Character vector or formula with the compartments in the
+##'     model to include in the plot. Default includes all
+##'     compartments in the model. Can also be a formula that
+##'     specifies the compartments that define the cases with a
+##'     disease or that have a specific characteristic (numerator),
+##'     and the compartments that define the entire population of
+##'     interest (denominator). The left-hand-side of the formula
+##'     defines the cases, and the right-hand-side defines the
+##'     population, for example, \code{I~S+I+R} in a \sQuote{SIR}
+##'     model (see \sQuote{Examples}). The \code{.}  (dot) is expanded
+##'     to all compartments, for example, \code{I~.}  is expanded to
+##'     \code{I~S+I+R} in a \sQuote{SIR} model (see
+##'     \sQuote{Examples}).
+##' @param level The level at which the prevalence is calculated at
+##'     each time point in \code{tspan}. 1 (population prevalence):
+##'     calculates the proportion of the individuals (cases) in the
+##'     population. 2 (node prevalence): calculates the proportion of
+##'     nodes with at least one case. 3 (within-node prevalence):
+##'     calculates the proportion of cases within each node. Default
+##'     is \code{1}.
+##' @param index Indices specifying the nodes to include when plotting
+##'     data. Plot one line for each node. Default (\code{index =
 ##'     NULL}) is to extract data from all nodes and plot the median
 ##'     count for the specified compartments.
 ##' @param range Show the quantile range of the count in each
@@ -226,126 +432,63 @@ init_plot_line_width <- function(lwd) {
 ##'
 ##' ## Plot the median and interquartile range of the  number
 ##' ## of infected individuals.
-##' plot(result, compartments = "I")
+##' plot(result, "I")
+##'
+##' ## Use the formula notation instead to plot the median and
+##' ## interquartile range of the number of infected individuals.
+##' plot(result, ~I)
 ##'
 ##' ## Plot the number of susceptible, infected
 ##' ## and recovered individuals in the first
 ##' ## three nodes.
-##' plot(result, node = 1:3, range = FALSE)
+##' plot(result, index = 1:3, range = FALSE)
 ##'
 ##' ## Use a stair step plot type instead.
-##' plot(result, node = 1:3, range = FALSE, type = "s")
+##' plot(result, index = 1:3, range = FALSE, type = "s")
 ##'
 ##' ## Plot the number of infected individuals in the first node.
-##' plot(result, compartments = "I", node = 1, range = FALSE)
+##' plot(result, "I", index = 1, range = FALSE)
+##'
+##' ## Plot the proportion of infected individuals (cases)
+##' ## in the population.
+##' plot(result, I~S+I+R)
+##'
+##' ## Plot the proportion of nodes with infected individuals.
+##' plot(result, I~S+I+R, level = 2)
+##'
+##' ## Plot the median and interquartile range of the proportion
+##' ## of infected individuals in each node
+##' plot(result, I~S+I+R, level = 3)
+##'
+##' ## Plot the proportion of infected individuals in the first
+##' ## three nodes.
+##' plot(result, I~S+I+R, level = 3, index = 1:3, range = FALSE)
 setMethod(
     "plot",
-    signature(x = "SimInf_model"),
-    function(x, compartments = NULL, node = NULL, range = 0.5, ...) {
-        if (identical(dim(x@U), c(0L, 0L))) {
-            stop("Please run the model first, the 'U' matrix is empty.",
-                 call. = FALSE)
-        }
+    signature(x = "SimInf_model", y = "ANY"),
+    function(x, y, level = 1, index = NULL, range = 0.5, ...) {
+        if (missing(y))
+            y <- NULL
 
-        argv <- list(...)
-
-        compartments <- init_plot_compartments(x, compartments)
-        node <- init_plot_node(x, node)
-
-        ## Create a matrix with one row for each line in the plot.
-        if (identical(range, FALSE)) {
-            ## Extract subset of data from U
-            i <- rep(compartments, length(node))
-            i <- i + rep((node - 1) * Nc(x), each = length(compartments))
-            m <- x@U[i, seq_len(ncol(x@U)), drop = FALSE]
+        if (isTRUE(compartments_has_lhs(y))) {
+            pd <- init_plot_prevalence_data(x, y, level, index, range)
         } else {
-            ## Check range argument
-            if (any(!is.numeric(range), !identical(length(range), 1L),
-                    range < 0, range > 1)) {
-                stop("'range' must be FALSE or a value between 0 and 1.",
-                     call. = FALSE)
-            }
-            range <- (1 - range) / 2
+            compartments <- match_compartments(compartments = y,
+                                               ok_combine = TRUE,
+                                               ok_lhs = FALSE,
+                                               U = rownames(x@S),
+                                               V = rownames(x@v0))
 
-            m <- matrix(0, nrow = length(compartments),
-                        ncol = length(x@tspan))
-
-            ## Matrices for quantile range
-            mu <- m
-            ml <- m
-
-            for (i in seq_len(length(compartments))) {
-                k <- seq(from = compartments[i], to = dim(x@U)[1],
-                         by = Nc(x))
-                u <- apply(x@U[k[node], seq_len(ncol(x@U)), drop = FALSE],
-                           2,
-                           quantile,
-                           probs = c(range, 0.5, 1 - range))
-                ml[i, ] <- u[1, ]
-                m[i, ] <- u[2, ]
-                mu[i, ] <- u[3, ]
-            }
-
-            range <- TRUE
+            pd <- init_plot_trajectory_data(x, compartments, index, range)
         }
 
-        lty <- init_plot_line_type(argv$lty, compartments, m)
-        col <- init_plot_color(argv$col, compartments, m)
-        argv$type <- init_plot_type(argv$type)
-        argv$lwd <- init_plot_line_width(argv$lwd)
+        argv <- init_plot_argv(x, y, pd, ...)
+        lty <- init_plot_line_type(argv$lty, pd$compartments, pd$each)
+        col <- init_plot_color(argv$col, pd$compartments, pd$each)
 
-        ## Settings for the y-axis.
-        argv$ylab <- "N"
-        if (isTRUE(range)) {
-            argv$ylim <- c(0, max(mu))
-        } else {
-            argv$ylim <- c(0, max(m))
-        }
+        plot_data(pd, argv, lty, col)
 
-        ## Settings for the x-axis
-        if (is.null(names(x@tspan))) {
-            xx <- x@tspan
-            argv$xlab <- "Time"
-        } else {
-            xx <- as.Date(names(x@tspan))
-            argv$xlab <- "Date"
-        }
-
-        savepar <- par(mar = c(2, 4, 1, 1), oma = c(4, 1, 0, 0),
-                       xpd = TRUE)
-        on.exit(par(savepar))
-
-        ## Plot lines
-        for (i in seq_len(dim(m)[1])) {
-            argv$x <- xx
-            argv$y <- m[i, ]
-            argv$col <- col[i]
-            argv$lty <- lty[i]
-            if (i == 1) {
-                do.call(plot, argv)
-                title(xlab = argv$xlab, outer = TRUE, line = 0)
-            } else {
-                do.call(lines, argv)
-            }
-            if (isTRUE(range)) {
-                polygon(x = c(xx, rev(xx)), y = c(mu[i, ], rev(ml[i, ])),
-                        col = adjustcolor(col[i], alpha.f = 0.1),
-                        border = NA)
-            }
-        }
-
-        ## Add the legend below plot. The default legend is the names
-        ## of the compartments.
-        if (is.null(argv$legend))
-            argv$legend <- rownames(x@S)[compartments]
-        par(fig = c(0, 1, 0, 1), oma = c(0, 0, 0, 0),
-            mar = c(0, 0, 0, 0), new = TRUE)
-        plot(0, 0, type = "n", bty = "n", xaxt = "n", yaxt = "n")
-        legend("bottom", inset = c(0, 0),
-               lty = lty[seq_len(length(compartments))],
-               col = col[seq_len(length(compartments))],
-               bty = "n", horiz = TRUE, legend = argv$legend,
-               lwd = argv$lwd)
+        invisible(NULL)
     }
 )